pyproject.toml
[tool.poetry]
name = "atomic-habits"
version = "0.1.0"
description = ""
authors = ["Anton Pashinov <troxin1988@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
redis = "^5.1.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = "migrations"
